Move iree/utils down a level to iree/tools/utils.
