+++ iree/tools/BUILD
@@ -25,7 +25,7 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/utils:vm_util",
+        "//runtime/src/iree/tools/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -45,8 +45,8 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/testing:benchmark",
-        "//runtime/src/iree/utils:trace_replay",
-        "//runtime/src/iree/utils:yaml_util",
+        "//runtime/src/iree/tools/utils:trace_replay",
+        "//runtime/src/iree/tools/utils:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
@@ -69,7 +69,7 @@ cc_binary(
         "//runtime/src/iree/modules/check",
         "//runtime/src/iree/modules/hal",
         "//runtime/src/iree/testing:gtest",
-        "//runtime/src/iree/utils:vm_util",
+        "//runtime/src/iree/tools/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
     ],
@@ -137,7 +137,7 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/utils:vm_util",
+        "//runtime/src/iree/tools/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -163,7 +163,7 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/utils:vm_util",
+        "//runtime/src/iree/tools/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -180,8 +180,8 @@ cc_binary(
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
-        "//runtime/src/iree/utils:trace_replay",
-        "//runtime/src/iree/utils:yaml_util",
+        "//runtime/src/iree/tools/utils:trace_replay",
+        "//runtime/src/iree/tools/utils:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
@@ -199,9 +199,9 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/utils:cpu_features",
-        "//runtime/src/iree/utils:trace_replay",
-        "//runtime/src/iree/utils:yaml_util",
+        "//runtime/src/iree/tools/utils:cpu_features",
+        "//runtime/src/iree/tools/utils:trace_replay",
+        "//runtime/src/iree/tools/utils:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
