+++ iree/tools/iree-run-module-main.cc
@@ -20,7 +20,7 @@
 #include "iree/hal/api.h"
 #include "iree/hal/drivers/init.h"
 #include "iree/modules/hal/module.h"
-#include "iree/utils/vm_util.h"
+#include "iree/tools/utils/vm_util.h"
 #include "iree/vm/api.h"
 #include "iree/vm/bytecode_module.h"
 #include "iree/vm/ref_cc.h"
