+++ iree/tools/iree-run-trace-main.c
@@ -13,8 +13,8 @@
 #include "iree/base/internal/flags.h"
 #include "iree/hal/api.h"
 #include "iree/hal/drivers/init.h"
-#include "iree/utils/trace_replay.h"
-#include "iree/utils/yaml_util.h"
+#include "iree/tools/utils/trace_replay.h"
+#include "iree/tools/utils/yaml_util.h"
 #include "iree/vm/api.h"
 
 IREE_FLAG(bool, trace_execution, false, "Traces VM execution to stderr.");
